tools: deploy_firmware: change shell.expect which checks command for ssh connection
We have a problem with broken logs when connect
to the board using UART. We use it once
when configuring vlans for ssh connection.
Sometimes shell.expect can be broken with
RDKB logs that spams a lot to the console.

Make shell.expect shorter so rdkb can't
paste their logs there.

PPM-1916

Signed-off-by: Dmytro Puz <d.puz@inango-systems.com>