agent: topology_task: report connected BSSID for backhaul STAs
The current value we use for "network_membership" for our backhaul
STAs is incorrect. We set it to the MAC address of the backhaul STA if
it's connected, while we should set it to the BSSID it's connected to.

Note that this is a quick fix for PPM-1892, as there are bigger issues
with our list of local interfaces that will be fixed in PPM-1943.

Fixes PPM-1892

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>