ci: trigger CGR builds in scheduled pipelines
Add a new build job for the CGR: build-for-cgr.

When started, it will create a "bridge job", that will trigger a new
pipeline in the CGR repository, and wait for it to complete (exactly
like it's currently the case for the "platform" bridge job, except
that in this case it's across two different Gitlab repositories). The
status of the job will also be reported.

Users with permissions to access the CGR private repository will be
able to click on the new pipeline, which will bring them to the
private repository to see the build result and artifacts. For other
users, only the status of the pipeline is available.

Note that for this to work, the user creating the pipeline (and so
also the job) has to have the permissions to create a new pipeline in
the CGR private repository. If it's not the case, the job will fail
with an error message.

A new conditional variable, BUILD_CGR, is also introduced because we
don't necessarily want to build for the CGR for *every* scheduled
pipelines. Builds will only start if BUILD_CGR is set to "yes". The
variable is also filled from our main pipeline, to make it available
in the "platform" pipeline.

Fixes PPM-1936

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>