ci: add stability tests for platform Boardfarm tests
We already have CI scheduled tests to increases the chances that we
notice an issue with the Docker Boardfarm tests.

Since we anyway mostly don't use the Boardfarm platform setup during
the night, do the same for the Boardfarm tests on NEC WX3000HP and
RDK-B Turris Omnia.

If the "STABILITY_TESTS" variable is set, 10 additional tests will be
run for each platform.

For it to work, the STABILITY_TESTS variable is also exported to the
platform pipeline.

Note that the rule we use for the variable in the platform pipeline is
different than the one in the main pipeline. This is because even if
the variable was not set in the main pipeline, it will be set in the
child pipeline (to the literal name of the variable in that
case). This is also explained in the description of [1] (case 4).

[1]: https://gitlab.com/gitlab-org/gitlab/-/issues/281201

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>