tlvf: Change how conditional parameter works
Currently, to add a conditional parameter use this syntax:

my_element:
  _condition:
    _field_name: my_presence_field
    _field_value: MY_ELEMENT_PRESENT

If the parameter under "_field_name" is equal to the value specified
under "_field_value", the condition is evaluated to 'true' and the
conditional parameter is added.

This implementation is very limiting since it only allows comparing
a single parameter to a single value.

On R3, we need to support a condition that compares a parameter to
multiple values. Such a thing is not supported by the current
implementation.
e.g. Add a BSSID field if some error code is equal to 0x07 or 0x08.

To support the requirement with maximum flexibility to the future Change
the implementation to the following way:

my_element:
  _condition:
    _using_members: member1 member2 member3 [...]
    _phrase: member1 == member2 && member3 > 3

Under "_using_members" need to specify which member are participating in
the condition phrase.
Under "_phrase" need to put any C++ syntax conditional phrase.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>