controller: nbapi: remove node usages for stations
In station dm setter methods, get_node and type CLIENT check is used.

BH Stations are not type of CLIENT, so it leads to
not update datamodel of these stations.

In this commit,

New database get_station method is used instead of get_node
Unnecessary log prints are removed for amx setters and ret_val check
is added.

Code is cleaned from repeated comments.

https://jira.prplfoundation.org/browse/PPM-1945
--
Signed-off-by: Cihangir Odabas <codabas@assia-inc.com>