tests: ClientSteeringDummy: configure SSIDs before connecting STA
The client steering dummy test connects STAs, but does not configure
SSIDs. This is probably a remnant from the time where the SSIDs were
kept between the different tests.

As can be seen in PPM-1949, not configuring the SSIDs mean the
controller database doesn't contain any BSS. Once we switch to using
the new node type this will be a problem, so fix it now.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>