common: son_wireless_utils: remove non-ascii character
In commit 1528164cc40758365d96f8ec2f9b69228dca5725, a non-ascii
character was inserted (probably from a copy-paste).

Instead of being a regular dash, this character is a Unicode "EM
Dash" ([1]).

In some cases (if the default encoding used is not UTF-8?),
enumAutoPrint will fail with:
Traceback (most recent call last):
  File "framework/tlvf/../external/enumAutoPrint/enum_auto_print.py", line 289, in upgrade_enum
    lines = f.readlines()
  File "/usr/lib/python3.6/encodings/ascii.py", line 26, in decode
    return codecs.ascii_decode(input, self.errors)[0]
UnicodeDecodeError: 'ascii' codec can't decode byte 0xe2 in position 3862: ordinal not in range(128)

Since we don't need this character to be a UTF-8 character anyway,
just replace it with the ascii dash.

[1]: https://www.fileformat.info/info/unicode/char/2014/index.htm

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>