tools: deploy_firmware.py:  split device-specific parts to tools/device
This commit is a mechanical change of moving most of the content of
deploy_firmware.py to a new folder, "tools/device".

Later on we will introduce separate scripts to deploy firmwares and
configure devices, where each script re-uses some part of the current
deploy_firmware.py script. Do the split now to make it easier later.

Note that this commit breaks the existing script, which will only be
fixes in a follow-up commit to make it easier to review.

In addition to splitting into different files, the following changes
are introduced:

- Add __init__.py file to make the device folder a package.
- Add the license header to every file while at it.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>