Add pyproject.toml
Since we're dealing with python files in the prplMesh repository, add
a pyproject.toml file which allows configuring multiple python-related
tools.

For now, it only includes configuration for two of them: pylint and
isort.

isort is used to sort imports automatically. The configuration skips
files ignored by Git.
It also adds headings for the different sorted sections, to make it
clearer to the reader.

pylint can be used to lint files to make sure they don't contain
syntax errors for example. The current configuration disables the
"convention" and "refactor" messages (as we currently have many of
those), to keep only the warnings and errors.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>