Check subband only for 5G BSS/Radios
Getting rid of misleading unnecessary error logs.
When the channel is 0 or 1, it is fraught with
irrelevant errors in the logs. Since the functions
which_freq, which_subband and some others only work
for 5G.
Add check to understand is it 5G or not.
If 5G then check subband.

PPM-1944

Signed-off-by: Dmytro Puz <d.puz@inango-systems.com>