tools: device: add SerialDevice
To be able to manipulate serial shells using pexpect easily, a
SerialDevice class is introduced.

In follow-up commits, other parts manipulating serial devices will be
refactored to make use of it.

Note that this script is literally imported from the easymesh_cert
repository. Unfortunately the easymesh_cert repository is only
available for certification tests, so we need a copy of the module in
the prplMesh repository.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>