tools: device: serial: add expect_prompt_on_connect
Expecting the prompt on connect is generally useful, because it allows
to empty the pexpect buffer to start from a clean state.

In some cases however it can be needed to not do that (e.g. because
the device is in the middle of a reboot when we connect to the serial
device).

Add a new optional parameter to control whether to expect the prompt
on connect or not (it defaults to True to keep the same behavior as
before by default).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>