ci: platform: add test-flashing-on-device tests
Currently the upgrading images is only tested when the images we build
change.

To make it easier to test the upgrade mechanism, introduce a manual
test that does the upgrade on the 4 devices we currently have in the
Essensium Boardfarm testbed.

A condition is also added so that if the pipeline is for a merge
request which has the "test-upgrade-image" label, the platform
pipeline as well as the new manual job will be run automatically.

Note that the RDK-B Turris Omnia has its own script section since it
has a different way to use deploy_firmware.py.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>