.gitlab-ci: use reference tags for expensive-job
When the expensive-job anchor was introduced, Gitlab didn't provide a
builtin way to re-use rules across multiple jobs.

Now Gitlab has introduced reference tags ([1]), which allows to do the
same in a builtin way. It's more flexible, as it will allow us to
extend the expensive-job rules with additional rules in the places
where it's used (see [2]).

[1]: https://docs.gitlab.com/ee/ci/yaml/yaml_optimization.html#reference-tags
[2]: https://docs.gitlab.com/ee/ci/jobs/job_control.html#reuse-rules-in-different-jobs

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>