tools: device: don't set custom prompts
In some cases we used to set a custom prompt to something easier to
"expect".

In the future, we will also need to handle u-boot prompts, which we
cannot change without re-compiling anyway.

To make future changes easier, do not set a custom prompt, and instead
expect the default one on prplOS.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>