tools: device: TurrisRdkb: make sure prplOS is running before upgrading
Currently the sysupgrade method (which in this case is used to flash
RDK-B manually) doesn't check what OS is currently loaded.

If the currently loaded OS is RDK-B, it will happily try to mount the
in-use rootfs and remove everything from it (leading to a device that
can only be recovered by power cycling it).

To make sure this cannot happen, the following changes are introduced:

- Move is_prplwrt_ready to be an instance method. We'll use it in
other places, and there is no real reason why it should be an inner
function.

- In sysupgrade(), make sure we load prplOS if it's not the case
yet, and wait for it to be loaded.

- To make it less likely that we try to remove the running rootfs if
burn_rdkb_on_board() is called independently in the future, also check
the running OS before anything else. If it's RDK-B, abort.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>