tools: device: TurrisRdkb: increase reboot time to 180 seconds
The default pexpect timeout is 30 seconds, which really a small amount
of time for the device to shutdown RDK-B, and boot prplOS in all
cases.

To give it more time to do so, increase the timeout to 180 seconds.

Note that this is likely to be the issue reported in PPM-1914.

Fixes PPM-1914.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>