agent: Prevent error logs in controller upon CAC completion
When a ZWDFS radio interface exist, Upon CAC completion, AP manager
sends a CAC completion notification the agent, then the agent
notifies the controller about the CAC completion.
When the controller receives this notifications
(both from the regular radio and from the ZWDFS radio),
it tries to get a node from its DB based on the radio MAC.
Since the controller does not familiar with ZWDFS radio interface,
the controller will print an error log.

PPM-1960

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>