optimal_path_task: increase RSSI measurement timeout
With the increased measurement window poll rate we see that the optimal
path reaches a timeout scenario for the RSSI measurements.
Need to increase the timeout of the optimal path RSSI measurements to
10 seconds to reflect this change, and avoid the timeout scenario.

This commit provides a quick solution to thee issue described in
PPM-1967.
Once PPM-1967 is resolved need to revert this change.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>