tools: device: check_serial_type: silence kernel messages
The current check_serial_type implementation is very fragile: it only
reads 25 lines from the device in order to get the OS type.

If the kernel outputs messages, the 25 lines might not be enough to
get the line we expect.

The implementation could be changed to be more robust in the future,
but for now just silence any kernel output before checking the content
of '/etc/os-release'.

Fixes:
- https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/2175371488
- https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/2175371469

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>