PPM-1924: configure devices from CI
This is a follow-up of: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/merge_requests/2898

In this MR:

- Some refactoring is done to break the TurrisRdkb's inheritance from
PrplOSDevice.
- Some more refactoring is done to have ways to reboot devices. The
possibility to stop in the bootloader is also added.
- A new class is added for each device. The per-device classes are
meant to hold device-specific things like the bootloader prompt, the
key sequence that allows to stop the bootloader, the way to upgrade
through the booltloader, etc.
- Configuration files for the CI devices are added.
- The configuration of each device is done when upgrading images.
- A standalone script to configure devices is also added.
- Some smaller things are fixed in `tools/device` and
`tools/deploy_firmware.py`.

As it could bring its own set of issues, the prplOS 1.4 integration
itself will be done in a separate MR.

Once this MR is merged, it should be safe to work on prplOS 1.4 images
independently (as when each MRs run, the devices will be re-configured
again, which will allow the tests to run).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
(cherry picked from 34b8880aa65285ee765ccf48232d0abb2768b0af^..ff46b951d2201ce8717699269cc2e86fd9258463)