tools: rename prplWrt to prplOS
The prplWrt project has been renamed to prplOS a while ago, so rename
it in our device scripts.

- The file for the prplwrt module is renamed to "prplos".

- The prplwrtDevice class is renamed "prplOSDevice" and users of the
class are adapted.

- In docstrings, mentions of prplWrt are renamed to "prplOS".

- In ShellType, PRPLWRT is renamed to PRPLOS.

Note that the usage as well as any mention of the
"/etc/prplwrt-version" file are not updated as part of this
commit. Since it's a file, we need to rename it in the prplOS builds
first.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>