tools: device: introduce GenericDevice class
Currently, the TurrisRdkb class inherits from PrplOSDevice.

This is unfortunate:

- RDK-B has nothing to do with prplOS, and features we would like to
implement in the PrplOSDevice class would be automatically be
inherited by the TurrisRdkb. This makes it difficult to modify
PrplOSDevice with enough confidence that nothing will break in
TurrisRdkb.

- One of the benefits of inheritance is code de-duplication, where
common parts can be re-used. The TurrisRdkb class doesn't take
advantage of it, as it re-defines almost every part that it
inherits (making the inheritance useless).

In other words, we need to break this inheritance.

This commit introduces a new class: GenericDevice. It will be used to
holds parts that are common to every device (including prplOS and
RDK-B-based devices).

At the same time, the parts of prplOSDevice that were common to every
device (even RDK-B) are moved to GenericDevice. The docstrings are
also adapted in some places (to remove references to prplOS).

The parts of prplOSDevice that are specific to prplOS are moved to
GenericPrplOS.

In addition, upgrade_uboot() is renamed to upgrade_bootloader() (as
some RDK-B devices, like the Cougar Run, don't use u-boot but can
still be upgraded through their bootloader).

With all of this, we'll be able to make the TurrisRdkb class inherit
from GenericDevice, and not from anything prplOS-specific anymore.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>