tools: device: add stop_in_bootloader()
Add an optional argument that, when set to true, makes reboot() stop
the device in its bootloader. This will be needed to upgrade devices
from their bootloaders.

We also need to introduce variables for bootloader prompts,
expressions and sequences, since those can vary from one bootloader to
an other.

The default values are for the u-boot version of the Turris Omnia, but
they can be overridden per-device (like its the case for the Axepoint
in this commit).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>