son_slave: move DB actions on client notifications
We see that, before the controller connection is established, clients
that are already connected are not reported as part of the topology.
This is because we do not add the client to the agent DB's
associated_clients.

On client associated and disconnected notifications need to update the
agent DB before checking if the agent is connected to a controller.

Signed-off-by: Itay Elenzweig <ielenzweig@maxlinear.com>