tools: device: axepoint: move upgrade_bootloader() to GenericPrplOS
The prplOS-common parts are moved to GenericPrplOS.

The device-specific parts are moved into a new method,
upgrade_from_u_boot(), that will be overridden in
subclasses (currently only for the axepoint).

In GenericPrplOS, upgrade_bootloader() is also slightly modified to
make use of the new reboot() method.

The part that killed the "S99prplmesh" init is removed. As we will
soon configure the devices from the prplMesh repository, it's much
less likely that this will still be needed.

Removing parts of the overlay, which was specific to the axepoint, is
also removed. It's not so important to keep it at this stage, since
we'll anyway wipe the overlay at every upgrade for every device in the
near future.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>