tools: device: add support for rebooting unknown Linux devices
When the Turris Omnia is running in "rescue shell" mode, there is no
os-release that we can use to differentiate the OS.

Still, we need to be able to tell if it's currently stopped in its
bootloader, or running the "rescue shell".

Add a new ShellType: LINUX_UNKNOWN. It will be used in case we could
tell the device is running Linux, but not exactly which OS.

check_serial_type() is changed to use the new LINUX_UNKNOWN. Instead
of always reading 25 lines of serial output (which hangs if the device
is running Linux but doesn't have /etc/os-release, like the Turris
Omnia rescue shell), we now use pexpect itself to look for OSTYPE_RE.
If it was found, we search for the OS type like we used to do.
If it was not found, we use dmesg to try to guess if we're running
Linux.
If we get the "Usage: dmesg" output, then we know we're running
LINUX_UNKNOWN.

This new logic is now used by reboot(), to also use the "reboot"
command for LINUX_UNKNOWN.

While at it, the reboot command itself is also improved. When a device
is running prplOS for example, in some rare cases it might not be able
to reboot using "reboot" alone (when the device is in a broken,
state). In those cases, "reboot -f" allows to skip some of the
shutdown procedure and proceed with the reboot. Furthermore, the
Turris Omnia rescue shell doesn't reboot at all with the "reboot"
command alone.
The command now uses "reboot", gives the device 15 seconds to reboot,
and then use "reboot -f" if needed.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>