tools: device: add TurrisPrplOS
Add a new class for a Turris Omnia running prplOS.

It overrides upgrade_from_u_boot() to upgrade the device through the
bootloader.

It provides two way to do the upgrade. The first one is by loading the
image from u-boot using tftp, then write it to the emmc.

On some devices (some more recent versions of the Turris Omnia),
writing to the emmc fails:
---
mmc write ${kernel_addr_r} 0 ${blkcnt}
MMC write: dev # 0, block # 0, count 62274 ... mmc write failed
0 blocks written: ERROR
---

In the prplOS repository, a u-boot patch ([1]) is used to enable
writing to the emmc and overcome this problem. If possible we would
prefer to avoid having to build and flash a custom u-boot version. In
the meantime the patch also unfortunately received negative feedback.

To avoid this u-boot bug entirely, another mechanism for the upgade is
used. The Turris Omnia has two flash memories: the NAND one where the
main OS is stored, and a NOR one which among other things hosts a
rescue image (which gives access to a rescue shell). This means that
even if we screw up the main OS, we can reasonably assume that the
rescue shell is still available on the device we need to flash. This
rescue shell is used by upgrade_from_u_boot() to load the new image
over SSH, then write it to the emmc.

The upgrade_from_u_boot() method will first attempt a regular write
from u-boot, and if its fails fall back to flashing from the rescue
shell.

The main deploy_firmware.py script is also adapted to create the
correct TurrisRdkb or TurrisPrplOS object. --os-type is not
needed (because the same information is already carried in
"args.device") so remove it.

Note that currently deploy_firmware.py will use sysupgrade for
TurrisPrplOS as that's still the default. In other words,
upgrade_from_u_boot() is not in use yet.

[1]: https://lists.denx.de/pipermail/u-boot/2021-February/441964.html

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>