ci: configuration: add CI devices configurations
These configurations will be used in a follow-up commit, to configure
the devices after they have been flashed.

Note that for the GL.iNet used in Boardfarm we have to specifically
enable the radios, while we don't do it for the prplOS Turris Omnia.

This is because the Turris Omnia already has a wireless configuration
in 'tests/certification/turris-omnia/' that is applied automatically.

Because the number of changes is already high, moving this wireless
configuration to the new configuration folder is left for a follow-up
merge request.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>