tools: device: add get_device
In follow-up commits, we'll want to construct devices outside of
deploy_firmware.py. As we don't want to duplicate the logic of
constructing the correct device, move it to a separate function.

Note that this new function cannot be in the "generic" module, as it
also contains the base class for every device (it would create
circular imports). For the same reason it also cannot be in the
"utils" module.

While we're at it, also remove the "kernel" argument for
TurrisRdkb. The kernel argument is always the same (it's even
statically written as "zImage" on the device). With the removal of
this argument, the constructor for this device has the same signature
as the others.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>