controller: topology response: fix wrong assume of radio from bss
According to standard;

A Multi-AP Agent shall indicate each BSS that is operating in
PWR_ON or PWR_SAVE mode as a separate (802.11)
Local Interface in the Device Information Type TLV
in the 1905 Topology Response message.

This interface list contains wired and wireless interfaces.
For wireless interfaces, BH and FH BSSs are reported.
Also, BH STA interface is reported with different role.

Removed code acts as Radio is reported in that list instead of BSS.
And not reported radios are removed from the database.

With these changes, if a device has different a RUID than FH BSSID,
the radio will not be removed accidentally anymore.

https://jira.prplfoundation.org/browse/PPM-1114
--
Signed-off-by: Cihangir Odabas <codabas@assia-inc.com>