tools: add a script to configure devices
Add a new module: configuration, that makes it easier to configure a
Generic device over a serial line.
The script writes a shell configuration file to the device, then
executes it.

Because serial lines can be unreliable, an md5 hash of the
configuration file is compared to make it slightly more robust,

A new standalone script, tools/configure_device.py is also added. It
allows to make use of the new module to configure a device from the
CLI.

In a follow-up commit, the new module will also be used to optionally
configure devices from deploy_firmware.py.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>