tools: deploy_firmware: add support for configuration files
As part of PPM-1924, we want to get rid of the device settings when we
flash a new image. For this to be possible, we have to apply the
device configuration from this repository.

Add a new command line argument to deploy_firmware.py, to allow the CI
to configure devices before it tries to use them.

Also remove the reachable check from GenericPrplOS, since it's now
done by deploy_firmware.py (after applying the configuration).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>