agent: Fix process termination when ZWDFS wlan interface exist
The agent process is terminating because of unsupported hostap interface type
of one of the radio interface.
This is caused because in some platform we have 4 radio interfaces; 2.4ghz, 5ghz,
6ghz and ZWDFS interface, but in beerocks_agent.conf we have 3 agent iterface sections.
The missing interface is filled with empty values, resulting in the termination of the
agent process.

Related commits for more explanation:
* 9508f6d9 - agent: remove hostap_iface from beerock_agent.conf
* 48e8b685 - agent: expand IRE_MAX_SLAVES to support tri-band

PPM-1970

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>