tools: deploy_firmware.py: upgrade device if needs_upgrade() fails
If needs_upgrade() cannot determine if the device has to be upgraded,
it could be for multiple reasons. For example, the device might be
stuck in its bootloader. It could also simply not be reachable over
SSH anymore (as needs_upgrade() currently uses an SSH connection).
In any of those cases, just do the upgrade anyway, to try to recover
the device automatically.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>