ci: configure devices after upgrades
Since with the prplOS 1.4 upgrade the device settings will be lost, we
need to configure them whenever we do an upgrade.

Use the new '--configuration' command line parameter of
deploy_firmware.py to configure the devices.

Note that since we don't strictly need it for the prplOS 1.4 upgrade,
the RDK-B Turris Omnia is left aside for now. It needs to be tested
separately later on.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>