Transport: DeDuplicationKeyCompare: Discard duplicated Topology Notification
Relayed multicast messages are sent both as multicast and as unicast
messages, and it's up to the receiver to discard the duplicated frame
based on the MID. However, the current de-duplication implementation
uses the destination address as a criteria to decide whether a frame is
a duplicate or not. For the relayed multicast frames, the destination
address is different (multicast vs unicast) but the frame should still
be considered as a duplicate, to avoid processing the same info twice.

When checking duplication for "reliable multicast frames", consider both
the unicast and the multicast versions.

Signed-off-by: Sylvain LE ROUX <prplmesh.opensources@sagemcom.com>