framework: tlvf: add length test when conditional elements are present
To make sure the length is computed correctly, add a new test to check
it. The new test creates a ERROR_RESPONSE_MESSAGE messages, containing
a tlvProfile2ErrorCode TLV.

The initial length of the TLV is checked. Then, a conditional
element (a MAC address in this case) is added, and the length is
checked again.

While at it, also add length checks to two other tests that related to
conditional elements (even though they were not affected by the bug
that was fixed in the previous commit).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>