tools: device: TurrisRdkb: upgrade device even if new version is older
As explained in PPM-1988, the new image is not flashed if it's older
than the one the device is currently running.

In CI, we want to test the version that was built in the same
pipeline, no matter if it was built before or after the one currently
on the device.

Fixes PPM-1988

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>