ci: configuration: port uci.sh configuration file
We currently have two ways to configure devices: using UCI scripts in
tests/certification, and using the new configuration files in
ci/configuration.

Since ci/configuration is the new way to configure CI devices, move
the wireless configuration there, with the aim of removing the ones in
tests/certification entirely.

There are only some small modifications to accommodate shellcheck, and
to commit uci changes and restart the network only once.

While at it, a reference to PPM-1928 is also added.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>