tools: deploy_firmware.py: always apply the configuration if there is one
Up to now, the configuration was only applied if the device also had
to be upgraded.

Change it to always apply the configuration if there is one, even if
no upgrade was needed.

We also only wait for the initialization time if a configuration was
applied or the device was upgraded.

Checking if the device is reachable over SSH at the end is done in all
cases.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>