common: Add 'program' parameter validation
We have a validation function for the 'program' parameter.
This function compares the value of the program parameter to hardcoded
strings, which are all lower case.

The UCC send the 'dev_exec_action' command with the 'program' parameter
in uppercase.
Convert the program parameter to lower case before comparing it.

Add a 'program' parameter validation to the 'dev_exec_action' command
handler.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>