agent: topology task: check every interface list to identify old device
When TOPOLOGY_DISCOVERY_MESSAGE is received,
reported neighbor is searched over list.

If device is new reported device,
topology notification is send.

There was a bug while iterating over interface list,
it is fixed with this commit.

Device is searched over all interfaces,
interfaces are cleared from device if interface is not received iface.

If received interface does not contain device/neighbor,
we set this as new device and send notification.

https://jira.prplfoundation.org/browse/PPM-1997
--
Signed-off-by: Cihangir Odabas <codabas@assia-inc.com>