tools: device: TurrisPrplOS: add a delay before flashing commands
As described in PPM-1990, in some cases the serial commands get
scrambled when we try to flash a black Turris Omnia (the old
revision).

It's clear that we don't wait for long enough after sending the
interrupt character. Even clearing the pexpect buffer before sending
the interrupt and expecting the prompt again doesn't work though: it
helps, but every now and then the commands still get scrambled.

Add a one second delay after sending the interrupt. This is not the
most elegant way to fix it, but it was tested manually and seems to
fix the issue. For manual validation, the device was flashed 10 times
starting from u-boot and 10 times starting from the OS, and it
succeeded every time.

Fixes PPM-1990

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>