tools: device: TurrisRdkb increase delay after reboot
As can be seen in [1], flashing the RDK-B device still occasionally
fails because the shell is not yet ready when we send the first
commands.

Commit 1021cab7c8671727a4a396d1926d4dbfd5980ef3 fixed a similar issue
when booting RDK-B, but there remain one when we boot prplOS.

Since is_prplos_ready() is used when we reboot into prplOS, increase
the delay that is already there.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/2258606416

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>