cmake: add feature flag at compilation
Up until now, there were some features that were compiled
only on RDKB platforms. However, there is a requirement to
compile those features on additional platforms.
Thus, A feature flag is created, so the feature compilation
will be based on feature flags instead of platform flags.
In case one wants to add feature flag in the future, he should append
it to FEATURE list, and to CMakeList.txt under files directory.
The next commit will include the renaming of ifdef compilation
flags in the source and header files.

Signed-off-by: Yoni Abecassis <yabecassis@maxlinear.com>