tools: device: Axepoint: send dummy command after flashing
As can be seen in [1], currently after the first successful flash, the
another attempt is made.

This is because the last command that was executed is executed again
when an empty line is sent (for the reboot).

Add a dummy command at the end, so that in case it gets repeated,
nothing happens.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/2264975677

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>