tools: device: send commands char by char with a delay
With the current implementation, characters would often get lost when
sending them, which would making the flashing fail.

See for example the following jobs, which fail in different ways, but
all due to missing characters in the commands sent over serial:

- https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/2265491697
- https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/2265512002
- https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/2264525850

A possible explanation is that we send commands too fast, and the
serial devices are sometimes unable to cope with so many characters at
the same time.

To give every device more time to process the characters we send, send
the characters one by one, with a delay between them.

To make this transition easy, SerialDevice inherits from fdspawn,
which means we can transparently use it as our pexpect shell. The
users of the class are adapted to use the SerialDevice directly,
instead of its old .shell member.

In SerialDevice, sendline() is overridden to send the commands
character by character.

Also reduce the default delay between characters to 10
milliseconds. As we also configure the devices over the serial line,
we don't want it to be too slow.

Part of this commit was generated using the following snippet:

for i in tools/device/*.py ; do sed -ri '/shell = ser.shell/d' "$i" ; sed -ri 's/(.*\) )as ser:/\1as shell:/g' "$i" ; done

Fixes PPM-2012

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>