tools: device: generic: do not force reboot if not needed
As can be seen for example in [1], some specific Axepoint devices have
troubles with the reboot command: sometimes, they would not boot
anymore and stay stuck at the u-boot stage until they are power
cycled.

Since we didn't have this issue before, it's possibly caused by the
'reboot -f' command, which bypasses part of the shutdown procedure.

The command as it was previously used always triggered 'reboot -f' to
run. Change it to make sure it's only run after some delay (in case
the device refuses to reboot without the '-f' argument).

Fixes PPM-2011

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/2268395492

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>