tlvf: Add support of conditional variables on bitfield member
Preparative commit.
This MR will add in one of the future messages a conditional varaible
which is based on a member of a bitfield. This is not supported since
there is a problem with the generated code which create wrong C++ syntax
so this is not supported.
Add the reqired changes in the tlvf script so it will be supported.

When acessing an element of struct through a pointer, need to change
the access operator from '.' to '->'.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>