tlvf: Remove WSC namespace from AttrList
AttrList is a generic class template. It is not related to the WSC
standard.
Remove the WSC namespace from the file, and update every file that
uses it.
On that opportunity, fix the include guard name.

As part of this commit, move the WscAttrList.h file to the WSC folder
since it is not generic.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>