tools: device: Axepoint: remove overlay before upgrading
For every other platform, we remove the overlay before upgrading (to
make sure that a safe, default configuration is generated again).

To be consistent with the other platforms, remove the overlay for
Axepoint-based devices as well.

This also uncovered some differences with the new default
configuration compared to the old one:

- Some interfaces are added to the guest bridge. Make them part of the
LAN bridge instead (PPM-2019).

- By default there is no backhaul STAs anymore, so add them (one per
radio).

- The dummy interface (default_radio100 and default_radio102) now have
'start_disabled=1'. Set it to 0.

Fixes PPM-2022

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>