ci: certification: fix device configuration file for R2
The '.run-certification-tests' template job is used both for R1 and R2
tests.

When configuring the devices using the '--configuration' parameter of
deploy_firmware.py, we need to take the configuration file that
corresponds to the correct revision.

Replace "R1" with "$EASYMESH_REVISION" so that the correct revision is
used for R2 as well.

Fixes PPM-2028

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>