ci: configuration: Axepoint: do not fail if syslog-ng cannot reload
On Axepoint-based devices, syslog-ng takes more time to start. If we
try to reload it before it has started, it leads to the following
error (see for example [1]):
Error connecting control socket, socket='/var/syslog-ng.ctl', error='No such file or directory'

In that case, we can ignore the error because when syslog-ng starts,
it will anyway re-create the /var/log/messages file if needed.

Note that even though the Axepoint and the NEC WX3000HP are supposed
to be almost identical, we haven't seen this issue on the NEC
WX3000HP.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/2303244140

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>